Add New Directory
Files and directories placed in the new `include` directory are copied
to the root of the destination ISO file if the `-i, --include` option
is given. An example being a post-installation configuration script.

This provides a convenient way of copying these files and directories from
the USB flash drive to a freshly installed Ubuntu Desktop either as part of
the preseed - see `example.seed` - or manually post-installation.